:Move avail-subxt examples to e2e tests
